* rename sds-node-agent to storage-network-controller
